v1_requirements.txt
alpaca-trade-api==3.1.1
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
lightgbm==4.5.0
scipy==1.14.1
pyyaml==6.0.2
websocket-client==1.8.0
orjson==3.10.6
pydantic==2.9.2
loguru==0.7.2
apscheduler==3.10.4
uvloop==0.20.0; sys_platform != "win32"
fastapi==0.115.5
uvicorn==0.32.0
prometheus-client==0.21.0